Import button on history screen now loads excel file and parses it into the history, replacing the old one. Imported games always show the results screen when activating a player.
It's very basic, no checks or balances are done (yet?) but it appears to work. Just don't feed it dashboard versions because it will attempt to turn everything into (empty) history elements and it will take forever. Only feed it raw or sanitized versions.

Also note that the export log doesn't contain everything the history elements normally contain, so some things are missing. We'll have to wait and see what the effects of that are going to be.