Fixed the existing results stuff. Moved all the duplicate stuff into the root component, and passed on the relevant values to the child components instead. It's a lot, but it is better for performance and code quality. The root component now contains a fuckton of computed properties, of which some could possibly be merged into each other for better performance and less boilerplate, BUT while it was written while I was low on focus this way it's at least possible to hook other results calculations in at various levels of granularity. One of them being the bar chart that's still coming up.
A new child component was added for the player properties (the average KPI). The bar chart will be another component, so that way each part will be a component.

Note that I currently have no way to varify that the calculations are correct other than believing they are: I do not have Excel and the online viewer can't deal with the SUMIFS, so I cannot compare values. Additionally, the imported history does not contain any prognosis adjustments due to an old bug I fixed a few commits ago, so that is not reflected in the (imported) results either.