Non-accidentally fixed the "deleteRecord" bug, by rearranging the promise logic. Also fixed the "End game" error by properly resetting the rounds. The global trading schedule is now part of the game configuration, and is automatically loaded and used. Default is set to 3-5 minutes, since I dunno.
I found out what caused the errors because the global trading schedule needed to be set on "Next round" and not on "End game". I wrapped it in an if, and then realized that a lot of other stuff also had to go inside this if. It was trying to set things that did not exist onto things that did not exist when the game was over.